renamed math module
